Adding some comments
