+++ src/ir/gates/parameterized/mprz.rs
@@ -1,25 +1,13 @@
-use crate::ir::gates::utils::{rot_z, get_indices};
+use crate::ir::gates::utils::{rot_z, get_indices, svd};
 use crate::ir::gates::{Gradient, Size};
 use crate::ir::gates::{Optimize, Unitary};
 
 use ndarray::{Array2, Array3, ArrayViewMut2};
 use ndarray_linalg::c64;
-use ndarray_linalg::SVD;
 use crate::squaremat::Matmul;
 
-fn svd(matrix: ArrayViewMut2<c64>) -> (Array2<c64>, Array2<c64>) {
-    let result = matrix.svd(true, true);
-    let actual_result = match result {
-        Ok(res)  => res,
-        Err(_res) => panic!("Problem svding the matrix: {:?}", matrix),
-    };
-
-    // Safety: u/vt are the same size since matrix is a square matrix with sides of size `size`
-    (actual_result.0.unwrap(), actual_result.2.unwrap())
-}
-
-
-/// A gate representing a multiplexed Z rotation one 1 qubit
+/// A gate representing a multiplexed Z rotation with a target qudit
+///  and n -1 select qudits
 #[derive(Copy, Clone, Debug, PartialEq, Default)]
 pub struct MPRZGate {
     size: usize,
@@ -50,6 +38,10 @@ impl Unitary for MPRZGate {
     }
 
     fn get_utry(&self, _params: &[f64], _constant_gates: &[Array2<c64>]) -> Array2<c64> {
+        // Return the unitary matrix for the MPRZ gate
+        // If the target qudit is the bottom most qubit, then the matrix
+        // is block diagonal of RY rotations
+        // Otherwise, use the get_indices to permute the matrix appropriately
         let mut arr: Array2<c64> = Array2::zeros((self.dim, self.dim));
         let mut i: usize = 0;
         for param in _params {
@@ -68,6 +60,8 @@ impl Unitary for MPRZGate {
 
 impl Gradient for MPRZGate {
     fn get_grad(&self, _params: &[f64], _const_gates: &[Array2<c64>]) -> Array3<c64> {
+        // Each theta is independent, so the gradient is simply
+        // the derivative of the rotation matrix with respect to theta
         let mut grad: Array3<c64> = Array3::zeros((_params.len(), self.dim, self.dim));
 
         for (i, &param) in _params.iter().enumerate() {
@@ -103,6 +97,9 @@ impl Optimize for MPRZGate {
     fn optimize(&self, env_matrix: ArrayViewMut2<c64>) -> Vec<f64> {
         let mut thetas: Vec<f64> = Vec::new();
         let mut i: usize = 0;
+        // Each variable is independent, so you can optimize each RZ separately
+        // Simply take the angles of the environment matrix and subtract them
+        // to get the closest theta.
         while i < self.dim {
             let (x1, x2) = get_indices(i, self.target_qudit, self.size);
             let real = env_matrix[[x2, x2]].re;
